requirement.txt
tensorflow>=2.10.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.1.0
Pillow>=8.3.0
opencv-python>=4.5.0
scikit-image>=0.19.0
rasterio>=1.3.0
tqdm>=4.64.0
###Evzalutation
laspy>=2.0.0
cupy-cuda11x>=10.0.0

